ama-backend/requirements.txt
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# OpenAI API for real ChatGPT integration
openai==1.3.7

# Google GenerativeAI for real Gemini integration
google-generativeai==0.3.2

# Anthropic Claude for real strategic coordination
anthropic==0.7.8

# MCP (Model Context Protocol) dependencies
mcp==1.0.0
mcp-client==1.0.0

# Firestore and Google Cloud dependencies
google-cloud-firestore==2.13.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Agent coordination and communication
websockets==12.0
aiohttp==3.9.1
asyncio-mqtt==0.16.1

# Biofield validation and HRV processing
numpy==1.24.3
scipy==1.11.4
pandas==2.1.4
scikit-learn==1.3.2
pyhrv==0.4.4
biosppy==0.1.2

# Data processing and validation
jsonschema==4.20.0
cerberus==1.3.5

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Security and authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Environment and configuration
python-dotenv==1.0.0
click==8.1.7

# Additional utilities
python-dateutil==2.8.2
pytz==2023.3 